refactor: ISR for detail pages, separate server/client env, retry webhook
