+++ lib/api/server.ts
@@ -1,12 +1,18 @@
 /**
  * 서버 컴포넌트 전용 fetch 함수
  *
- * - Next.js native fetch + next: { tags } 로 ISR 캐싱
- * - 클라이언트에서 사용하는 lib/api/projects.ts (Axios) 와 별개
+ * - API_URL: 서버 전용 환경변수 (클라이언트 번들에 노출되지 않음)
+ * - NEXT_PUBLIC_API_URL: 클라이언트 Axios용 (브라우저에서 직접 호출)
  * - 이 파일은 'use client' 컴포넌트에서 import 금지
  */
 
-const API_URL = process.env.NEXT_PUBLIC_API_URL ?? 'https://158.180.75.205';
+// 서버에서만 실행되므로 NEXT_PUBLIC_ 없이 사용
+// Vercel 환경변수에 API_URL (서버 전용)을 별도로 추가해야 함
+// 없으면 NEXT_PUBLIC_API_URL fallback (기존 동작 유지)
+const API_URL =
+  process.env.API_URL ??
+  process.env.NEXT_PUBLIC_API_URL ??
+  'https://hsm9411.duckdns.org';
 
 // ─── Projects ────────────────────────────────────────────
